refactor(codegen): carve cross-node serialization codec out of llvm.rs
Relocate the CBOR / wire-text serialize and deserialize thunk emitters
(the `__hew_cbor_serialize_*` / `__hew_cbor_deserialize_*` C-ABI thunk
pair emission, the wire-text descriptor builders, and their layout-walk
helpers) from llvm.rs into a new `crate::wire` concern module.

Pure relocation: the module-build path still drives these emitters via
re-imported entry points, so every emitted `.ll` is byte-identical
(ll-diff verified byte-identical native + wasm32). No IR change.
