+++ hew-codegen-rs/src/lib.rs
@@ -39,6 +39,7 @@ pub(crate) mod runtime_abi;
 mod runtime_family_parity;
 pub(crate) mod suspend;
 pub(crate) mod thunks;
+pub(crate) mod wire;
 
 pub use llvm::{
     emit_module, emit_module_objects, validate_codegen_front, verify_pipeline, CodegenError,
